backend/requirements.txt
flask==1.1.2
pandas
matplotlib
tqdm

gunicorn
pytz

seaborn
gspread
oauth2client
Flask-Login
Flask-SQLAlchemy==2.5.1
Flask-Cors
python-dotenv
PyJWT
bcrypt
pymongo[snappy,gssapi,srv,tls]
